+++ R/local_dependence.R
@@ -44,11 +44,13 @@
 #' @references
 #' Yen, W. M. (1984). Effects of local item dependence on the fit and
 #' equating performance of the three-parameter logistic model.
-#' *Applied Psychological Measurement*, 8(2), 125–145.
+#' *Applied Psychological Measurement, 8*(2), 125--145.
+#' \doi{10.1177/014662168400800201}
 #'
 #' Christensen, K. B., Makransky, G., & Horton, M. (2017). Critical values
 #' for Yen's Q3: Identification of local dependence in the Rasch model.
-#' *Applied Psychological Measurement*, 41(3), 178–194.
+#' *Applied Psychological Measurement, 41*(3), 178--194.
+#' \doi{10.1177/0146621616677520}
 #'
 #' @export
 #'