+++ R/score_se.R
@@ -170,8 +170,7 @@ RMscoreSE <- function(data,
       y = "Ordinal sum score"
     ) +
     ggplot2::theme_bw() +
-    ggplot2::theme(axis.title.x = ggplot2::element_text(margin = ggplot2::margin(t = 12)),
-                   axis.title.y = ggplot2::element_text(margin = ggplot2::margin(r = 12)))
+    er2_axis_margins()
 }
 
 # ---------------------------------------------------------------------------